NR-10857 : created new component for animation of button text
